MVP/requirements.txt
google-search-results
playwright
pandas
python-dotenv
